First version added to GitHub
This program is currently broken.
The structure of the code is as follows:
-Ask the user for a dark matter mass (in GeV)
-Ask the user for how many random samples they would like
-If the variable 'autolimit' is set to "ask", ask the user if they would
like to limit the velocities that are created from the random samples.
(If "autolimit" is set to "yes" or "no", the code will automatically
limit or pass the function.)
-The limit function will identify which velocities in the random array
are outside of the limits and will create another random velocity in its
place. This will loop until either all velocities are within the limits
or until a "replacement limit" is reached.
-Plot a theoretical curve of the energy distribution
-Plot the randomly sampled velocities on the theoretical curve
(velocities are sampled from maxwellian and lognormal PDFs

There are also some lines that are commented out (denoted by ##) which
were designed to evaluate the area underneath the theoretical energy
distribution (for demonstrating normalisation).