dndE_exp - experimental version of main code (dndE.py) -- DM_VDF - short code to experiment with scipy_stats 'rv_continuous' classes.
dndE_exp:
In this version of the experimental build, the current output displays a
theoretical curve for a maxwellian probability distribution function
(PDF), as well as a random sampling of maxwellian numbers (amount chosen
by the user) and a variation of maxwellian numbers that have been
affected by noise.

DM_VDF:
This program is specifically to design a class structure for plotting a
specific dark matter maxwellian function. Currently, the program can
successfully plot a theoretical PDF curve, and a random sample of
maxwellian numbers to fit the curve. Still working on the 'rvs' function
for the class.