dndE_v3 fix. rv_continuous high processing time for 'rvs' function
After implementing most of vel_dist.py into dndE_v3.py, the result is a
working program, however not as intended. It was discovered that
utilising the rvs command with my_dmdv causes a tremendous amount of
processing time. 2 samples usually takes ~7 seconds to process, 10
samples is ~20 seconds and anything beyond 100 samples is unfathomable.
Also worth mentioning is the problem with passing units in and out of
my_dmdv; somehow the output of the function strips the values of their
units.